feat: Add main window with label and button
This commit adds the main window to the application, which includes a QLabel widget displaying "Hello!" and a QPushButton widget labeled "Press me!". The window is set to a minimum size of 400x300 pixels and a maximum size of 800x600 pixels.

The purpose of this commit is to provide a basic user interface for the application.